% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{distr}
\alias{distr}
\title{Distribution function}
\usage{
distr(x, dist, param, type = "d", model = NULL, ...)
}
\arguments{
\item{x}{data}

\item{dist}{distribution of interest}

\item{param}{parameters of the distribution function}

\item{type}{= 'd'}

\item{model}{= NULL model being fitted}

\item{...}{additional arguments to be passed to the distribution function}
}
\description{
Distribution function
}
