% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histline.R
\name{histline}
\alias{histline}
\title{makes histograms in line form}
\usage{
histline(
  height,
  breaks,
  lineonly = FALSE,
  outline = FALSE,
  fill = FALSE,
  ylim = range(height),
  xlab = "x",
  ylab = "y",
  ...
)
}
\arguments{
\item{height}{vector of bar heights, and breaks vector of cutpoints}

\item{breaks}{vector of cutpoints}

\item{lineonly}{if TRUE, uses lines() to plot histogram, else uses plot()}

\item{outline}{if TRUE, only outline of histogram is plotted}

\item{fill}{if TRUE, uses polygon() to fill bars}

\item{ylim}{y-axis limits}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{additional arguments passed to plot() or lines() if lineonly=TRUE, or polygon() if fill=TRUE
histogram from them using the function plot() (if lineonly==FALSE) or lines()
(if lineonly==TRUE). 
If fill==TRUE, uses polygon() to fill bars
If fill==TRUE, valid arguments to plot() or lines() are passed via argument(s) "..."
If outline==TRUE, only outline of histogram is plotted
If fill!=TRUE, valid arguments to polygon() are passed via argument(s) "..."}
}
\description{
Takes bar heights (height) and cutbpoints (breaks), and constructs a line-only
}
