% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{nupoint_krill}
\alias{nupoint_krill}
\title{Example data from a parallel density gradient: a multibeam survey of krill}
\format{
A data frame with 251 rows and 7 variables:
\describe{
  \item{transect}{Transect number}
  \item{x}{Shoal cross-track distance, m}
  \item{y}{Shoal depth, m}
  \item{r}{Shoal radial distance, m}
  \item{theta}{Shoal angle, radians}
  \item{bio.g}{Shoal biomass, g}
  \item{z}{Seabed depth under shoal, m.}
}
}
\source{
For data collection details see: Cox et al (2009) Multibeam echosounder observations reveal interactions between Antarctic krill and air-breathing predators. *Marine Ecology Progress Series*, 378, 199–209. Acknowledgements: Multi-beam instrument was loaned by J. Condiotty of Simrad USA. Data were collected in association with an NSF-funded (grant #06-OPP-33939) investigation of the Livingston Island nearshore environment. Support provided by US Antarctic Marine Living Resources Program, and engineering by Sea Technology Services.
}
\description{
Krill data set illustrating a parallel density gradient example. Observations of krill swarms made using a multibeam echosounder.
}
\keyword{datasets}
